+++ src/pages/Event/AllEvent.css
@@ -1,5 +1,4 @@
 .event {
-    background-color: white;
     height: 300px;
     border-radius: 15px;
     background: linear-gradient(rgba(13, 13, 14, 0.5), rgba(28, 27, 27, 0.5));
